Almost implemented writing sample + restructurized
