+++ ClickHaskell/ClickHaskell/NativeProtocol.hs
@@ -1,48 +1,38 @@
-{-# LANGUAGE 
+{-# LANGUAGE
     AllowAmbiguousTypes
   , DataKinds
-  , DefaultSignatures
   , DeriveAnyClass
   , DeriveGeneric
   , DerivingStrategies
   , DuplicateRecordFields
-  , GeneralizedNewtypeDeriving
   , LambdaCase
   , NamedFieldPuns
   , OverloadedStrings
-  , TemplateHaskell
   , UndecidableInstances
 #-}
 
-module ClickHaskell.NativeProtocol where
+module ClickHaskell.NativeProtocol
+  ( module ClickHaskell.NativeProtocol
+  , module ClickHaskell.NativeProtocol.Serialization
+  , module ClickHaskell.NativeProtocol.Versioning
+  ) where
 
 -- Internal dependencies
 import ClickHaskell.DbTypes
-import ClickHaskell.Tables ()
-import Paths_ClickHaskell (version)
+import ClickHaskell.NativeProtocol.Serialization
+import ClickHaskell.NativeProtocol.Versioning
 
 -- GHC included
 import Control.Exception (Exception, throw)
 import Data.Binary.Get
-import Data.Binary.Get.Internal (readN)
-import Data.Bits (Bits (..))
-import Data.ByteString as BS (length, take, StrictByteString)
 import Data.ByteString.Builder (Builder)
-import Data.ByteString.Builder as BS
-  ( byteString
-  , int16LE, int32LE, int64LE, int8
-  , word16LE, word32LE, word64LE, word8
-  )
 import Data.ByteString.Lazy.Internal as BL (ByteString(..))
 import Data.ByteString.Lazy.Char8 as BSL8 (head)
 import Data.ByteString.Lazy (LazyByteString)
 import Data.Char (ord)
 import Data.Typeable (Proxy (..))
-import Data.Version (Version (..), showVersion)
-import Data.WideWord (Int128 (..), Word128 (..))
 import GHC.Generics
-import GHC.TypeLits (KnownNat, Nat, natVal)
-import Language.Haskell.TH.Syntax (lift)
+import GHC.TypeLits (KnownNat, natVal)
 import Data.Text (Text)
 
 -- External
@@ -76,6 +66,22 @@ determineServerPacket sock = do
     if headByte <= fromEnum (maxBound :: ServerPacketType)
     then toEnum headByte
     else throw $ ProtocolImplementationError UnknownPacketType
+
+
+-- ** Bufferized reading
+
+bufferizedRead :: forall packet . Deserializable packet => ProtocolRevision -> IO LazyByteString -> IO packet
+bufferizedRead rev bufferFiller = runBufferReader bufferFiller (runGetIncremental (deserialize @packet rev)) BL.Empty
+
+runBufferReader :: Deserializable packet => IO LazyByteString -> Decoder packet -> LazyByteString -> IO packet
+runBufferReader bufferFiller (Partial decoder) (BL.Chunk bs mChunk)
+  = runBufferReader bufferFiller (decoder $ Just bs) mChunk
+runBufferReader bufferFiller (Partial decoder) BL.Empty = do
+  bufferFiller >>= \case
+    BL.Empty -> fail "Expected more bytes while reading packet" -- ToDo: Pass packet name
+    BL.Chunk bs mChunk -> runBufferReader bufferFiller (decoder $ Just bs) mChunk
+runBufferReader _bufferFiller (Done _leftover _consumed helloPacket) _input = pure helloPacket
+runBufferReader _bufferFiller (Fail _leftover _consumed msg) _currentBuffer = error msg
 
 
 -- ** HelloResponse
@@ -164,12 +170,6 @@ instance
 
 -- ** Hello
 
-data HelloParameters = MkHelloParameters
-  { chDatabase :: Text
-  , chLogin :: Text
-  , chPass :: Text
-  }
-
 mkHelloPacket :: HelloParameters -> HelloPacket
 mkHelloPacket MkHelloParameters{chDatabase, chLogin, chPass} =
   MkHelloPacket
@@ -182,7 +182,13 @@ mkHelloPacket MkHelloParameters{chDatabase, chLogin, chPass} =
     , user                 = toChType chLogin
     , password             = toChType chPass
     , adendum              = MkSinceRevision MkAdendum
-    } 
+    }
+
+data HelloParameters = MkHelloParameters
+  { chDatabase :: Text
+  , chLogin :: Text
+  , chPass :: Text
+  }
 
 data HelloPacket = MkHelloPacket
   { packet_type          :: Packet Hello
@@ -213,26 +219,6 @@ data PingPacket = MkPingPacket
 
 
 -- ** Query
-
-data QueryStage
-  = FetchColumns
-  | WithMergeableState
-  | Complete
-  | WithMergeableStateAfterAggregation
-  | WithMergeableStateAfterAggregationAndLimit
-  deriving (Enum)
-
-instance Serializable QueryStage where
-  serialize rev = serialize @UVarInt rev . fromIntegral . fromEnum
-
-queryStageCode :: QueryStage -> UVarInt
-queryStageCode = fromIntegral . fromEnum
-
-data Flags = IMPORTANT | CUSTOM | OBSOLETE
-flagCode :: Flags -> ChUInt64
-flagCode IMPORTANT = 0x01
-flagCode CUSTOM    = 0x02
-flagCode OBSOLETE  = 0x04
 
 mkQueryPacket :: ProtocolRevision -> ChString -> ChString -> QueryPacket
 mkQueryPacket chosenRev user query = MkQueryPacket
@@ -267,6 +253,26 @@ mkQueryPacket chosenRev user query = MkQueryPacket
   , parameters = MkSinceRevision MkQueryParameters
   }
 
+data QueryStage
+  = FetchColumns
+  | WithMergeableState
+  | Complete
+  | WithMergeableStateAfterAggregation
+  | WithMergeableStateAfterAggregationAndLimit
+  deriving (Enum)
+
+instance Serializable QueryStage where
+  serialize rev = serialize @UVarInt rev . fromIntegral . fromEnum
+
+queryStageCode :: QueryStage -> UVarInt
+queryStageCode = fromIntegral . fromEnum
+
+data Flags = IMPORTANT | CUSTOM | OBSOLETE
+flagCode :: Flags -> ChUInt64
+flagCode IMPORTANT = 0x01
+flagCode CUSTOM    = 0x02
+flagCode OBSOLETE  = 0x04
+
 data QueryPacket = MkQueryPacket
   { query_packet       :: Packet Query
   , query_id           :: ChString
@@ -295,7 +301,7 @@ data ClientInfo = MkClientInfo
   , client_revision              :: UVarInt
   , quota_key                    :: ChString `SinceRevision` DBMS_MIN_REVISION_WITH_QUOTA_KEY_IN_CLIENT_INFO
   , distrubuted_depth            :: UVarInt `SinceRevision` DBMS_MIN_PROTOCOL_VERSION_WITH_DISTRIBUTED_DEPTH
-  , client_patch                 :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_VERSION_PATCH 
+  , client_patch                 :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_VERSION_PATCH
   , open_telemetry               :: ChUInt8 `SinceRevision` DBMS_MIN_REVISION_WITH_OPENTELEMETRY
   , collaborate_with_initiator   :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_PARALLEL_REPLICAS
   , count_participating_replicas :: UVarInt `SinceRevision` DBMS_MIN_REVISION_WITH_PARALLEL_REPLICAS
@@ -343,8 +349,10 @@ data BlockInfo = MkBlockInfo
   }
   deriving (Generic, Serializable)
 
-mkDataPacket :: DataName -> DataPacket
-mkDataPacket dataName =
+type ColumnsCount = UVarInt
+type RowsCount = UVarInt
+mkDataPacket :: ColumnsCount -> RowsCount -> DataName -> DataPacket
+mkDataPacket columns rows dataName =
   MkDataPacket
     { packet_type   = MkPacket
     , data_name     = dataName
@@ -355,8 +363,8 @@ mkDataPacket dataName =
       , bucket_num   = -1
       , bi5          = 0
       }
-    , columns_count = 0
-    , rows_count    = 0
+    , columns_count = columns
+    , rows_count    = rows
     }
 
 
@@ -383,277 +391,3 @@ data ConnectionError
   = NoAdressResolved
   | EstablishTimeout
   deriving (Show, Exception)
-
-
-
-
--- * Serialization
-
--- ** Bufferized reading
-
-bufferizedRead :: forall packet . Deserializable packet => ProtocolRevision -> IO LazyByteString -> IO packet
-bufferizedRead rev bufferFiller = runBufferReader bufferFiller (runGetIncremental (deserialize @packet rev)) BL.Empty
-
-runBufferReader :: Deserializable packet => IO LazyByteString -> Decoder packet -> LazyByteString -> IO packet
-runBufferReader bufferFiller (Partial decoder) (BL.Chunk bs mChunk)
-  = runBufferReader bufferFiller (decoder $ Just bs) mChunk
-runBufferReader bufferFiller (Partial decoder) BL.Empty = do
-  bufferFiller >>= \case
-    BL.Empty -> fail "Expected more bytes while reading packet" -- ToDo: Pass packet name
-    BL.Chunk bs mChunk -> runBufferReader bufferFiller (decoder $ Just bs) mChunk
-runBufferReader _bufferFiller (Done _leftover _consumed helloPacket) _input = pure helloPacket
-runBufferReader _bufferFiller (Fail _leftover _consumed msg) _currentBuffer = error msg
-
-
--- ** Serializable
-
-class Serializable chType
-  where
-  default serialize :: (Generic chType, GSerializable (Rep chType)) => ProtocolRevision -> chType -> Builder
-  serialize :: ProtocolRevision -> chType -> Builder
-  serialize rev = gSerialize rev . from
-
-instance Serializable UVarInt where
-  serialize _ = go
-    where
-    go i
-      | i < 0x80 = word8 (fromIntegral i)
-      | otherwise = word8 (setBit (fromIntegral i) 7) <> go (unsafeShiftR i 7)
-instance Serializable ChUInt8 where serialize _ = word8 . fromChType
-instance Serializable ChUInt16 where serialize _ = word16LE . fromChType
-instance Serializable ChUInt32 where serialize _ = word32LE . fromChType
-instance Serializable ChUInt64 where serialize _ = word64LE . fromChType
-instance Serializable ChUInt128 where serialize _ = (\(Word128 hi lo) -> word64LE hi <> word64LE lo) . fromChType
-instance Serializable ChInt8 where serialize _ = int8 . fromChType
-instance Serializable ChInt16 where serialize _ = int16LE . fromChType
-instance Serializable ChInt32 where serialize _ = int32LE . fromChType
-instance Serializable ChInt64 where serialize _ = int64LE . fromChType
-instance Serializable ChInt128 where serialize _ = (\(Int128 hi lo) -> word64LE hi <> word64LE lo) . fromChType
-instance Serializable ChString where
-  serialize revision str
-    =  (serialize @UVarInt revision . fromIntegral . BS.length . fromChType) str
-    <> (BS.byteString . fromChType @_ @StrictByteString) str
-
-class GSerializable f
-  where
-  gSerialize :: ProtocolRevision -> f p -> Builder
-
-instance
-  GSerializable f
-  =>
-  GSerializable (D1 c (C1 c2 f))
-  where
-  gSerialize rev (M1 (M1 re)) = gSerialize rev re
-
-instance
-  GSerializable (left1 :*: (left2 :*: right))
-  =>
-  GSerializable ((left1 :*: left2) :*: right)
-  where
-  gSerialize rev ((left :*: left2) :*: right)
-    = gSerialize rev (left :*: (left2 :*: right))
-
-instance
-  Serializable chType
-  =>
-  GSerializable (S1 (MetaSel (Just typeName) a b f) (Rec0 chType))
-  where
-  gSerialize rev = serialize rev . unK1 . unM1
-
-instance
-  (Serializable chType, GSerializable right)
-  =>
-  GSerializable (S1 (MetaSel (Just typeName) a b f) (Rec0 chType) :*: right)
-  where
-  gSerialize rev (left :*: right)
-    = (serialize rev . unK1 . unM1 $ left) <> gSerialize rev right
-
-
--- ** Deserializable
-
-class
-  Deserializable chType
-  where
-  default deserialize :: (Generic chType, GDeserializable (Rep chType)) => ProtocolRevision -> Get chType
-  deserialize :: ProtocolRevision -> Get chType
-  deserialize rev = to <$> gDeserialize rev
-
-instance Deserializable UVarInt where
-  deserialize _ = go 0 (0 :: UVarInt)
-    where
-    go i o | i < 10 = do
-      byte <- getWord8
-      let o' = o .|. ((fromIntegral byte .&. 0x7f) `unsafeShiftL` (7 * i))
-      if byte .&. 0x80 == 0 then pure $! o' else go (i + 1) $! o'
-    go _ _ = fail "input exceeds varuint size"
-instance Deserializable ChUInt8 where deserialize _ = toChType <$> getWord8
-instance Deserializable ChUInt16 where deserialize _ = toChType <$> getWord16le
-instance Deserializable ChUInt32 where deserialize _ = toChType <$> getWord32le
-instance Deserializable ChUInt64 where deserialize _ = toChType <$> getWord64le
-instance Deserializable ChUInt128 where deserialize _ = toChType <$> (Word128 <$> getWord64le <*> getWord64le)
-instance Deserializable ChInt8 where deserialize _ = toChType <$> getInt8
-instance Deserializable ChInt16 where deserialize _ = toChType <$> getInt16le
-instance Deserializable ChInt32 where deserialize _ = toChType <$>  getInt32le
-instance Deserializable ChInt64 where deserialize _ = toChType <$>  getInt64le
-instance Deserializable ChInt128 where deserialize _ = toChType <$> (Int128 <$> getWord64le <*> getWord64le)
-instance Deserializable ChString where
-  deserialize revision = do 
-    strSize <- fromIntegral <$> deserialize @UVarInt revision
-    toChType <$> readN strSize (BS.take strSize)
-
-class GDeserializable f
-  where
-  gDeserialize :: ProtocolRevision -> Get (f p)
-
-instance
-  GDeserializable f
-  =>
-  GDeserializable (D1 c (C1 c2 f))
-  where
-  gDeserialize rev = M1 . M1 <$> gDeserialize rev
-
-instance
-  GDeserializable (left :*: (right1 :*: right2))
-  =>
-  GDeserializable ((left :*: right1) :*: right2)
-  where
-  gDeserialize rev =
-    (\(left :*: (right1 :*: right2)) -> (left :*: right1) :*: right2)
-    <$> gDeserialize rev
-
-instance
-  (Deserializable chType)
-  =>
-  GDeserializable (S1 (MetaSel (Just typeName) a b f) (Rec0 chType))
-  where
-  gDeserialize rev =  M1 . K1 <$> deserialize @chType rev
-
-instance
-  (Deserializable chType, GDeserializable right)
-  =>
-  GDeserializable (S1 (MetaSel (Just typeName) a b f) (Rec0 chType) :*: right)
-  where
-  gDeserialize rev = (:*:) <$> (M1 . K1 <$> deserialize @chType rev) <*> gDeserialize rev
-
-instance {-# OVERLAPPING #-}
-  GDeserializable right
-  =>
-  GDeserializable (S1 (MetaSel (Just "server_revision") a b f) (Rec0 UVarInt) :*: right)
-  where
-  gDeserialize rev = do
-    server_revision <- deserialize @UVarInt rev
-    (:*:) <$> (pure . M1 . K1 $ server_revision) <*> gDeserialize server_revision
-
-
-
-
--- * Versioning
-
--- ** Client versioning
-
-clientMajorVersion, clientMinorVersion, clientPatchVersion :: UVarInt
-clientMajorVersion = fromIntegral $(lift (versionBranch version !! 0))
-clientMinorVersion = fromIntegral $(lift (versionBranch version !! 1))
-clientPatchVersion = fromIntegral $(lift (versionBranch version !! 2))
-
-clientNameAndVersion :: ChString
-clientNameAndVersion = $(lift ("ClickHaskell-" <> showVersion version))
-
-
--- ** Protocol compatibility wrappers
-
-type ProtocolRevision = UVarInt
-
-afterRevision
-  :: forall (revision :: Nat) monoid
-  .  (KnownNat revision, Monoid monoid)
-  => ProtocolRevision -> monoid -> monoid
-afterRevision chosenRevision monoid =
-  if chosenRevision >= (fromIntegral . natVal) (Proxy @revision)
-  then monoid
-  else mempty
-
-latestSupportedRevision :: ProtocolRevision
-latestSupportedRevision = (fromIntegral . natVal) (Proxy @DBMS_TCP_PROTOCOL_VERSION)
-
-data SinceRevision a (revisionNumber :: Nat) = MkSinceRevision a | NotPresented
-
-instance
-  ( KnownNat revision
-  , Serializable chType
-  )
-  =>
-  Serializable (SinceRevision chType revision)
-  where
-  serialize rev (MkSinceRevision val) = afterRevision @revision rev (serialize rev val)
-  serialize rev NotPresented = afterRevision @revision rev (error "Unexpected error")
-
-instance
-  ( KnownNat revision
-  , Deserializable chType
-  )
-  =>
-  Deserializable (SinceRevision chType revision)
-  where
-  deserialize rev =
-    if rev >= (fromIntegral . natVal) (Proxy @revision)
-    then MkSinceRevision <$> deserialize @chType rev 
-    else pure NotPresented
-
-
--- ** Versions
-
-{-
-  Slightly modified C++ sources:
-  https://github.com/ClickHouse/ClickHouse/blob/eb4a74d7412a1fcf52727cd8b00b365d6b9ed86c/src/Core/ProtocolDefines.h#L6
--}
-type DBMS_TCP_PROTOCOL_VERSION = 54471;
-
-type DBMS_MIN_REVISION_WITH_CLIENT_INFO = 54032;
-type DBMS_MIN_REVISION_WITH_SERVER_TIMEZONE = 54058;
-type DBMS_MIN_REVISION_WITH_QUOTA_KEY_IN_CLIENT_INFO = 54060;
-type DBMS_MIN_REVISION_WITH_TABLES_STATUS = 54226;
-type DBMS_MIN_REVISION_WITH_TIME_ZONE_PARAMETER_IN_DATETIME_DATA_TYPE = 54337;
-type DBMS_MIN_REVISION_WITH_SERVER_DISPLAY_NAME = 54372;
-type DBMS_MIN_REVISION_WITH_VERSION_PATCH = 54401;
-type DBMS_MIN_REVISION_WITH_SERVER_LOGS = 54406;
-type DBMS_MIN_REVISION_WITH_CURRENT_AGGREGATION_VARIANT_SELECTION_METHOD = 54448;
-type DBMS_MIN_MAJOR_VERSION_WITH_CURRENT_AGGREGATION_VARIANT_SELECTION_METHOD = 21;
-type DBMS_MIN_MINOR_VERSION_WITH_CURRENT_AGGREGATION_VARIANT_SELECTION_METHOD = 4;
-type DBMS_MIN_REVISION_WITH_COLUMN_DEFAULTS_METADATA = 54410;
-type DBMS_MIN_REVISION_WITH_LOW_CARDINALITY_TYPE = 54405;
-type DBMS_MIN_REVISION_WITH_CLIENT_WRITE_INFO = 54420;
-type DBMS_MIN_REVISION_WITH_SETTINGS_SERIALIZED_AS_STRINGS = 54429;
-type DBMS_MIN_REVISION_WITH_SCALARS = 54429;
-type DBMS_MIN_REVISION_WITH_OPENTELEMETRY = 54442;
-type DBMS_MIN_REVISION_WITH_AGGREGATE_FUNCTIONS_VERSIONING = 54452;
-type DBMS_CLUSTER_PROCESSING_PROTOCOL_VERSION = 1;
-type DBMS_MIN_SUPPORTED_PARALLEL_REPLICAS_PROTOCOL_VERSION = 3;
-type DBMS_PARALLEL_REPLICAS_MIN_VERSION_WITH_MARK_SEGMENT_SIZE_FIELD = 4;
-type DBMS_PARALLEL_REPLICAS_PROTOCOL_VERSION = 4;
-type DBMS_MIN_REVISION_WITH_PARALLEL_REPLICAS = 54453;
-type DBMS_MERGE_TREE_PART_INFO_VERSION = 1;
-type DBMS_MIN_REVISION_WITH_INTERSERVER_SECRET = 54441;
-type DBMS_MIN_REVISION_WITH_X_FORWARDED_FOR_IN_CLIENT_INFO = 54443;
-type DBMS_MIN_REVISION_WITH_REFERER_IN_CLIENT_INFO = 54447;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_DISTRIBUTED_DEPTH = 54448;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_INCREMENTAL_PROFILE_EVENTS = 54451;
-type DBMS_MIN_REVISION_WITH_CUSTOM_SERIALIZATION = 54454;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_INITIAL_QUERY_START_TIME = 54449;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_PROFILE_EVENTS_IN_INSERT = 54456;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_VIEW_IF_PERMITTED = 54457;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_ADDENDUM = 54458;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_QUOTA_KEY = 54458;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_PARAMETERS = 54459;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_SERVER_QUERY_TIME_IN_PROGRESS = 54460;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_PASSWORD_COMPLEXITY_RULES = 54461;
-type DBMS_MIN_REVISION_WITH_INTERSERVER_SECRET_V2 = 54462;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_TOTAL_BYTES_IN_PROGRESS = 54463;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_TIMEZONE_UPDATES = 54464;
-type DBMS_MIN_REVISION_WITH_SPARSE_SERIALIZATION = 54465;
-type DBMS_MIN_REVISION_WITH_SSH_AUTHENTICATION = 54466;
-type DBMS_MIN_REVISION_WITH_TABLE_READ_ONLY_CHECK = 54467;
-type DBMS_MIN_REVISION_WITH_SYSTEM_KEYWORDS_TABLE = 54468;
-type DBMS_MIN_REVISION_WITH_ROWS_BEFORE_AGGREGATION = 54469;
-type DBMS_MIN_PROTOCOL_VERSION_WITH_CHUNKED_PACKETS = 54470;
-type DBMS_MIN_REVISION_WITH_VERSIONED_PARALLEL_REPLICAS_PROTOCOL = 54471;
